requirements.txt
moviepy
punctuator
Cython
braceexpand
editdistance
frozendict
inflect
kaldi-io
librosa
marshmallow
packaging
ruamel.yaml
soundfile
sox
unidecode
kaldi-python-io
kaldiio
scipy
pandas
g2p_en
pydub
pyannote.core
pyannote.metrics
pyamg
torch-stft
ipywidgets

